chore(deps): 更新 mingyang-admin-common依赖至 v0.0.5
- 升级 github.com/saas-mingyang/mingyang-admin-common从 v0.0.4 到 v0.0.5- 调整 ID 字段默认值生成方式，使用 sonyflake.NextID()
- 修改 Snowflake 相关方法返回类型从 int64

+++ orm/ent/mixins/id_int64.go
@@ -31,7 +31,7 @@ type IDInt64Mixin struct {
 
 func (IDInt64Mixin) Fields() []ent.Field {
 	return []ent.Field{
-		field.Int64("id").Default(sonyflake.NextID()),
+		field.Int64("id").Default(int64(sonyflake.NextID())),
 		field.Time("created_at").
 			Immutable().
 			Default(time.Now).
